+++ ios/IssueTracker/IssueTracker/Controller/SectionItemEditVC/Views/BVLabelViewCell.xib
@@ -17,13 +17,14 @@
                 <rect key="frame" x="0.0" y="0.0" width="394" height="64"/>
                 <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
                 <subviews>
-                    <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="tMJ-7W-fUB">
+                    <button opaque="NO" userInteractionEnabled="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="tMJ-7W-fUB">
                         <rect key="frame" x="354" y="21" width="20" height="22"/>
                         <constraints>
                             <constraint firstAttribute="width" constant="20" id="9u6-a3-0vo"/>
                             <constraint firstAttribute="height" constant="22" id="I1W-10-dzx"/>
                         </constraints>
                         <color key="tintColor" white="0.66666666666666663" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                        <state key="normal" image="sun.dust.fill" catalog="system"/>
                     </button>
                     <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="tke-fD-Avy">
                         <rect key="frame" x="20" y="10" width="100" height="44"/>
@@ -53,6 +54,7 @@
         </collectionViewCell>
     </objects>
     <resources>
+        <image name="sun.dust.fill" catalog="system" width="128" height="119"/>
         <systemColor name="systemBackgroundColor">
             <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
         </systemColor>
